[OCC] Make sure simulation mode is safe (no reboot, no grab) and works (SDL_ params)

+++ src/layout.py
@@ -436,5 +436,6 @@ class layout():
 	def reboot(self):
 		self.quit()
 		time.sleep(2)
-		os.system("reboot")
+		if not self.occ.simulate:
+			os.system("reboot")
